feat: Implemented the compositor
refactor: Major coding refactor due the composition system

refactor: code styling

+++ app/workspace.cpp
@@ -1,5 +1,5 @@
 #include <vector>
 #include "workspace.h"
 
-int current_workspace = 0;
+long unsigned int current_workspace = 0;
 std::vector<workspace> workspaces;
