+++ app/workspace.h
@@ -8,7 +8,7 @@ typedef struct workspace {
     std::vector<client> clients;
 } workspace;
 
-extern int current_workspace;
+extern long unsigned int current_workspace;
 extern std::vector<workspace> workspaces;
 
 #endif
